# UniversityStats/UniversityStats.Client/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UniversityStats.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/University": {
      "get": {
        "tags": [
          "University"
        ],
        "summary": "Return list of universities",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UniversityDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UniversityDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UniversityDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "University"
        ],
        "summary": "Post university's information to database",
        "requestBody": {
          "description": "University's information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UniversityDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UniversityDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UniversityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UniversityDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniversityDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniversityDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "University"
        ],
        "summary": "Correct university's information by id",
        "requestBody": {
          "description": "University's information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UniversityDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UniversityDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UniversityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UniversityDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniversityDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniversityDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/University/{registrationNumber}": {
      "get": {
        "tags": [
          "University"
        ],
        "summary": "Return university's information by registration number",
        "parameters": [
          {
            "name": "registrationNumber",
            "in": "path",
            "description": "University's registration number",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UniversityDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniversityDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniversityDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "University"
        ],
        "summary": "Delete university's information by registration number",
        "parameters": [
          {
            "name": "registrationNumber",
            "in": "path",
            "description": "Registration number",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "UniversityDto": {
        "type": "object",
        "properties": {
          "registrationNumber": {
            "type": "string",
            "description": "University's registration number",
            "nullable": true
          },
          "nameUniversity": {
            "type": "string",
            "description": "University's name",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "University's address",
            "nullable": true
          },
          "propertyType": {
            "type": "string",
            "description": "University's property type",
            "nullable": true
          },
          "buildingOwnership": {
            "type": "string",
            "description": "University's building owner",
            "nullable": true
          },
          "rectorFullName": {
            "type": "string",
            "description": "Rector's name",
            "nullable": true
          },
          "degree": {
            "type": "string",
            "description": "Rector's degree",
            "nullable": true
          },
          "title": {
            "type": "string",
            "description": "Rector's tittle",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Class for data transfer about university's information"
      }
    }
  }
}